chore: fix flapper
